requirements.txt
Flask==2.3.3
flask-cors==3.0.10
werkzeug==3.0.0
pycryptodome==3.19.0
numpy==1.26.4
qiskit==1.0.2
qiskit-aer==0.14.2
flask-limiter==3.5.0
cryptography==42.0.4
bcrypt==4.1.3
python-dotenv==1.0.1
